Move label_studio_to_bboxes to extraction_utilities.py
